@@ -1,6 +1,6 @@
 {
   "name": "polaris.shopify.com",
-  "version": "0.19.0",
+  "version": "0.21.1",
   "private": true,
   "scripts": {
     "build": "yarn gen-assets && next build",
@@ -20,19 +20,19 @@
   "dependencies": {
     "@floating-ui/react-dom-interactions": "^0.6.1",
     "@headlessui/react": "^1.6.5",
-    "@shopify/polaris": "^10.4.1",
+    "@shopify/polaris": "^10.7.0",
     "@shopify/polaris-icons": "^6.4.0",
-    "@shopify/polaris-tokens": "^6.1.0",
+    "@shopify/polaris-tokens": "^6.2.0",
     "codesandbox": "^2.2.3",
     "framer-motion": "^6.5.1",
     "fuse.js": "^6.5.3",
     "gray-matter": "^4.0.3",
     "lodash.get": "^4.4.2",
     "lodash.throttle": "^4.1.1",
-    "next": "12.1.0",
+    "next": "12.1.6",
     "prismjs": "^1.27.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-markdown": "^8.0.2",
     "remark-gfm": "^3.0.1",
     "use-dark-mode": "^2.3.1"
@@ -45,7 +45,7 @@
     "@types/marked": "^4.0.3",
     "@types/node": "17.0.21",
     "@types/prismjs": "^1.26.0",
-    "@types/react": "*",
+    "@types/react": "18.0.15",
     "chokidar": "^3.5.3",
     "eslint": "8.10.0",
     "eslint-config-next": "12.1.0",
