@@ -20,15 +20,18 @@ const getAppCode = (code: string) => {
 
 const indexCode = `
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import App from "./App";
 import { AppProvider } from "@shopify/polaris";
-import en from '@shopify/polaris/locales/en.json';
+import en from "@shopify/polaris/locales/en.json";
 import "@shopify/polaris/build/esm/styles.css";
-
-const app = <AppProvider i18n={en}><App /></AppProvider>;
-
-ReactDOM.render(app, document.getElementById("root"));
+const container = document.getElementById("root");
+const root = createRoot(container);
+root.render(
+  <AppProvider i18n={en}>
+    <App />
+  </AppProvider>
+);
 `;
 
 interface Props {
