@@ -1,10 +1,14 @@
 import type {FileInfo, API, Options} from 'jscodeshift';
 import postcss, {Plugin} from 'postcss';
 import valueParser from 'postcss-value-parser';
-import {toPx} from '@shopify/polaris-tokens';
 
 import {POLARIS_MIGRATOR_COMMENT} from '../../constants';
-import {hasNumericOperator} from '../../utilities/sass';
+import {
+  hasNumericOperator,
+  hasTransformableLength,
+  toTransformablePx,
+} from '../../utilities/sass';
+import {isKeyOf, isValueOf} from '../../utilities/type-guards';
 
 // List of the props we want to run this migration on
 const targetProps = [
@@ -32,11 +36,6 @@ const targetProps = [
   'margin-block-end',
 ] as const;
 
-type TargetProp = typeof targetProps[number];
-
-const isTargetProp = (propName: unknown): propName is TargetProp =>
-  targetProps.includes(propName as TargetProp);
-
 // Mapping of spacing tokens and their corresponding px values
 const spacingTokensMap = {
   '1px': '--p-space-025',
@@ -57,11 +56,6 @@ const spacingTokensMap = {
   '128px': '--p-space-32',
 } as const;
 
-type SpacingToken = keyof typeof spacingTokensMap;
-
-const isSpacingTokenValue = (value: unknown): value is SpacingToken =>
-  Object.keys(spacingTokensMap).includes(value as SpacingToken);
-
 const processed = Symbol('processed');
 
 /**
@@ -81,7 +75,7 @@ const plugin = (_options: PluginOptions = {}): Plugin => {
 
       const prop = decl.prop;
 
-      if (!isTargetProp(prop)) return;
+      if (!isValueOf(targetProps, prop)) return;
 
       const parsedValue = valueParser(decl.value);
 
@@ -91,15 +85,11 @@ const plugin = (_options: PluginOptions = {}): Plugin => {
         if (node.type === 'function') {
           return StopWalkingFunctionNodes;
         } else if (node.type === 'word') {
-          const dimension = valueParser.unit(node.value);
+          const valueInPx = toTransformablePx(node.value);
 
-          if (isTransformableLength(dimension)) {
-            const dimensionInPx = toPx(`${dimension.number}${dimension.unit}`);
+          if (!isKeyOf(spacingTokensMap, valueInPx)) return;
 
-            if (!isSpacingTokenValue(dimensionInPx)) return;
-
-            node.value = `var(${spacingTokensMap[dimensionInPx]})`;
-          }
+          node.value = `var(${spacingTokensMap[valueInPx]})`;
         }
       });
 
@@ -127,36 +117,4 @@ export default function replaceSassLengths(
   return postcss(plugin(options)).process(fileInfo.source, {
     syntax: require('postcss-scss'),
   }).css;
-}
-
-/**
- * All transformable dimension units. These values are used to determine
- * if a decl.value can be converted to pixels and mapped to a Polaris custom property.
- */
-const transformableLengthUnits = ['px', 'rem'];
-
-function isTransformableLength(
-  dimension: false | valueParser.Dimension,
-): dimension is valueParser.Dimension {
-  if (!dimension) return false;
-
-  // Zero is the only unitless length we can transform
-  if (dimension.unit === '' && dimension.number === '0') return true;
-
-  return transformableLengthUnits.includes(dimension.unit);
-}
-
-function hasTransformableLength(parsedValue: valueParser.ParsedValue): boolean {
-  let transformableLength = false;
-
-  parsedValue.walk((node) => {
-    if (
-      node.type === 'word' &&
-      isTransformableLength(valueParser.unit(node.value))
-    ) {
-      transformableLength = true;
-    }
-  });
-
-  return transformableLength;
 }
