@@ -76,14 +76,6 @@ export function DetailsPage() {
   const [supportSubject, setSupportSubject] = useState('');
   const [supportMessage, setSupportMessage] = useState('');
 
-  const handleSubjectChange = useCallback(
-    (value) => setSupportSubject(value),
-    [],
-  );
-  const handleMessageChange = useCallback(
-    (value) => setSupportMessage(value),
-    [],
-  );
   const handleDiscard = useCallback(() => {
     setEmailFieldValue(defaultState.current.emailFieldValue);
     setNameFieldValue(defaultState.current.nameFieldValue);
@@ -101,7 +93,7 @@ export function DetailsPage() {
     setSearchActive(false);
     setSearchValue('');
   }, []);
-  const handleSearchFieldChange = useCallback((value) => {
+  const handleSearchFieldChange = useCallback((value: string) => {
     setSearchValue(value);
     setSearchActive(value.length > 0);
   }, []);
@@ -477,15 +469,13 @@ export function DetailsPage() {
   // ---- Select ----
   const [selected, setSelected] = useState('today');
 
-  const handleSelectChange = useCallback((value) => setSelected(value), []);
-
   const options = [
     {label: 'Keyboard', value: 'keyboard'},
     {label: 'Accessories', value: 'accessories'},
     {label: 'Last 7 days', value: 'lastWeek'},
   ];
 
-  const handleChange = useCallback((newValue) => {
+  const handleChange = useCallback((newValue: string) => {
     setDescriptionValue(newValue);
     setPreviewValue(newValue);
   }, []);
@@ -633,14 +623,14 @@ export function DetailsPage() {
               <Select
                 label="Product type"
                 options={options}
-                onChange={handleSelectChange}
+                onChange={setSelected}
                 value={selected}
               />
               <br />
               <Select
                 label="Vendor"
                 options={options}
-                onChange={handleSelectChange}
+                onChange={setSelected}
                 value={selected}
               />
             </Card.Section>
@@ -686,13 +676,13 @@ export function DetailsPage() {
           <TextField
             label="Subject"
             value={supportSubject}
-            onChange={handleSubjectChange}
+            onChange={setSupportSubject}
             autoComplete="off"
           />
           <TextField
             label="Message"
             value={supportMessage}
-            onChange={handleMessageChange}
+            onChange={setSupportMessage}
             autoComplete="off"
             multiline
           />
