@@ -1,5 +1,4 @@
 import React, {ReactElement} from 'react';
-import type {ThenType} from '@shopify/useful-types';
 import {mountWithApp} from 'tests/utilities';
 import type {Element as ElementType} from '@shopify/react-testing';
 import {act} from 'react-dom/test-utils';
@@ -205,7 +204,7 @@ describe('<Checkbox />', () => {
 });
 
 function triggerCheckboxEvent(
-  checkbox: ThenType<ReturnType<typeof mountWithTable>>,
+  checkbox: Awaited<ReturnType<typeof mountWithTable>>,
   eventType: any,
   event: {[key: string]: any},
 ) {
