@@ -65,6 +65,7 @@ interface State {
 
 export class PopoverOverlay extends PureComponent<PopoverOverlayProps, State> {
   static contextType = PortalsManagerContext;
+  context!: React.ContextType<typeof PortalsManagerContext>;
 
   state: State = {
     transitionStatus: this.props.active
@@ -281,7 +282,7 @@ export class PopoverOverlay extends PureComponent<PopoverOverlayProps, State> {
     } = this;
     const composedPath = event.composedPath();
     const wasDescendant = preventCloseOnChildOverlayClick
-      ? wasPolarisPortalDescendant(composedPath, this.context.container)
+      ? wasPolarisPortalDescendant(composedPath, this.context!.container)
       : wasContentNodeDescendant(composedPath, contentNode);
     const isActivatorDescendant = nodeContainsDescendant(activator, target);
     if (
@@ -299,8 +300,19 @@ export class PopoverOverlay extends PureComponent<PopoverOverlayProps, State> {
     this.props.onClose(PopoverCloseSource.ScrollOut);
   };
 
-  private handleEscape = () => {
-    this.props.onClose(PopoverCloseSource.EscapeKeypress);
+  private handleEscape = (event: Event) => {
+    const target = event.target as HTMLElement;
+    const {
+      contentNode,
+      props: {activator},
+    } = this;
+    const composedPath = event.composedPath();
+    const wasDescendant = wasContentNodeDescendant(composedPath, contentNode);
+    const isActivatorDescendant = nodeContainsDescendant(activator, target);
+
+    if (wasDescendant || isActivatorDescendant) {
+      this.props.onClose(PopoverCloseSource.EscapeKeypress);
+    }
   };
 
   private handleFocusFirstItem = () => {
@@ -312,10 +324,7 @@ export class PopoverOverlay extends PureComponent<PopoverOverlayProps, State> {
   };
 }
 
-function renderPopoverContent(
-  children: React.ReactNode,
-  props?: Partial<PaneProps>,
-) {
+function renderPopoverContent(children: React.ReactNode, props: PaneProps) {
   const childrenArray = Children.toArray(children);
   if (isElementOfType(childrenArray[0], Pane)) {
     return childrenArray;
