@@ -3,10 +3,10 @@ import React, {Children, isValidElement} from 'react';
 // Wraps `element` in `Component`, if it is not already an instance of
 // `Component`. If `props` is passed, those will be added as props on the
 // wrapped component. If `element` is null, the component is not wrapped.
-export function wrapWithComponent<TProps>(
+export function wrapWithComponent<TProps extends React.PropsWithChildren>(
   element: React.ReactNode | null | undefined,
   Component: React.ComponentType<TProps>,
-  props: TProps,
+  props: TProps & JSX.IntrinsicAttributes,
 ): React.ReactNode {
   if (element == null) {
     return null;
